Separate dfu-related classes and actual program.
