+++ memory_map/stm32f407.json
@@ -1,0 +1,50 @@
+[
+  {
+    "address": "0x8000000",
+    "size": "0x4000"
+  },
+  {
+    "address": "0x8004000",
+    "size": "0x4000"
+  },
+  {
+    "address": "0x8008000",
+    "size": "0x4000"
+  },
+  {
+    "address": "0x800c000",
+    "size": "0x4000"
+  },
+  {
+    "address": "0x8010000",
+    "size": "0x10000"
+  },
+  {
+    "address": "0x8020000",
+    "size": "0x20000"
+  },
+  {
+    "address": "0x8040000",
+    "size": "0x20000"
+  },
+  {
+    "address": "0x8060000",
+    "size": "0x20000"
+  },
+  {
+    "address": "0x8080000",
+    "size": "0x20000"
+  },
+  {
+    "address": "0x80a0000",
+    "size": "0x20000"
+  },
+  {
+    "address": "0x80c0000",
+    "size": "0x20000"
+  },
+  {
+    "address": "0x80e0000",
+    "size": "0x20000"
+  }
+]
